more hand decryption related tests

+++ docs/abi.js
@@ -132,7 +132,7 @@ function HashDecryptCard(salt, table_id, handnum, encrypted) {
   for (let card = 0; card <= 51; card++) {
     // Generate the keccak256 hash of the concatenated salt, table_id, handnum, and card
     // console.log([salt, table_id, handnum, card])
-    const hash = ethers.keccak256(ethers.solidityPacked(['uint256', 'uint256', 'uint256', 'uint256'], [salt, table_id, handnum, card]));
+    const hash = ethers.keccak256(ethers.solidityPacked(['uint256', 'uint256', 'uint256', 'uint8'], [salt, table_id, handnum, card]));
 
     // Check if the generated hash matches the encrypted data
     if (hash === ethers.hexlify(encrypted)) {
